refactored code to new service
